packages/aipype/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aipype"
version = "0.1.0a2"
description = "A modular AI agent framework with declarative pipeline-based task orchestration"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Pritam Barhate", email = "pritambarhate@mobisoftinfotech.com"},
]
keywords = ["ai", "agents", "llm", "pipeline", "framework", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

# Core dependencies only - lean package
dependencies = [
    "litellm>=1.77.1",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "types-requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "readability-lxml>=0.8.1",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "chardet>=5.0.0",
    "moviepy>=1.0.3",
    "openai>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/mobisoftinfotech/aipype"
Repository = "https://github.com/mobisoftinfotech/aipype"
Issues = "https://github.com/mobisoftinfotech/aipype/issues"


[tool.pytest.ini_options]
testpaths = ["tests", "integration_tests"]
markers = [
    "integration: marks tests as integration tests",
    "slow: marks tests that take significant time to run",
]
filterwarnings = [
    "ignore:There is no current event loop:DeprecationWarning:litellm",
    "ignore:Pydantic serializer warnings:UserWarning:pydantic",
    "ignore:Use 'content=<.*>' to upload raw bytes/text content:DeprecationWarning:httpx",
]
